break out proposal header links
