+++ client/scripts/views/pages/view_proposal/header.tsx
@@ -1,31 +1,20 @@
-/* eslint-disable @typescript-eslint/ban-types */
+/* @jsx m */
+
 import m from 'mithril';
+
 import { Button, Tag, MenuItem, Input } from 'construct-ui';
 
 import app from 'state';
 import { navigateToSubpage } from 'app';
 import { slugify } from 'utils';
 
-import {
-  pluralize,
-  link,
-  externalLink,
-  extractDomain,
-  offchainThreadStageToLabel,
-} from 'helpers';
+import { pluralize, link, offchainThreadStageToLabel } from 'helpers';
 import { getProposalUrlPath, proposalSlugToFriendlyName } from 'identifiers';
-import {
-  OffchainThread,
-  OffchainThreadKind,
-  OffchainThreadStage,
-  AnyProposal,
-} from 'models';
-import { ProposalType } from 'types';
+import { OffchainThread, OffchainThreadStage, AnyProposal } from 'models';
 
 import { notifySuccess } from 'controllers/app/notifications';
 
 import { confirmationModalWithText } from 'views/modals/confirm_modal';
-import { SnapshotProposal } from 'client/scripts/helpers/snapshot_utils';
 import { activeQuillEditorHasText, GlobalStatus } from './body';
 import { IProposalPageState } from '.';
 import { CWIcon } from '../../components/component_kit/cw_icons/cw_icon';
@@ -34,112 +23,8 @@ import {
   getStatusText,
 } from '../../components/proposal_card/helpers';
 
-export const ProposalHeaderExternalLink: m.Component<{
-  proposal: AnyProposal | OffchainThread;
-}> = {
-  view: (vnode) => {
-    const { proposal } = vnode.attrs;
-    if (!proposal) return;
-    if (!(proposal instanceof OffchainThread)) return;
-    if (proposal.kind !== OffchainThreadKind.Link) return;
-    return m('.ProposalHeaderExternalLink', [
-      externalLink('a.external-link', proposal.url, [
-        extractDomain(proposal.url),
-        m(CWIcon, { iconName: 'externalLink' }),
-      ]),
-    ]);
-  },
-};
-
-export const ProposalHeaderBlockExplorerLink: m.Component<{
-  proposal: AnyProposal;
-}> = {
-  view: (vnode) => {
-    const { proposal } = vnode.attrs;
-    if (!proposal || !proposal['blockExplorerLink']) return;
-    return m('.ProposalHeaderBlockExplorerLink', [
-      externalLink('a.voting-link', proposal['blockExplorerLink'], [
-        proposal['blockExplorerLinkLabel'] ||
-          extractDomain(proposal['blockExplorerLink']),
-        m(CWIcon, { iconName: 'externalLink' }),
-      ]),
-    ]);
-  },
-};
-
-export const ProposalHeaderExternalSnapshotLink: m.Component<{
-  proposal: SnapshotProposal;
-  spaceId: string;
-}> = {
-  view: (vnode) => {
-    const { proposal, spaceId } = vnode.attrs;
-    if (!proposal || !proposal.id || !spaceId) return;
-
-    return m('.ProposalHeaderBlockExplorerLink', [
-      externalLink(
-        'a.voting-link',
-        `https://snapshot.org/#/${spaceId}/proposal/${proposal.id}`,
-        [`View on Snapshot`, m(CWIcon, { iconName: 'externalLink' })]
-      ),
-    ]);
-  },
-};
-
-export const ProposalHeaderVotingInterfaceLink: m.Component<{
-  proposal: AnyProposal;
-}> = {
-  view: (vnode) => {
-    const { proposal } = vnode.attrs;
-    if (!proposal || !proposal['votingInterfaceLink']) return;
-    return m('.ProposalHeaderVotingInterfaceLink', [
-      externalLink('a.voting-link', proposal['votingInterfaceLink'], [
-        proposal['votingInterfaceLinkLabel'] ||
-          extractDomain(proposal['votingInterfaceLink']),
-        m(CWIcon, { iconName: 'externalLink', iconSize: 'small' }),
-      ]),
-    ]);
-  },
-};
-
-export const ProposalHeaderThreadLink: m.Component<{ proposal: AnyProposal }> =
-  {
-    view: (vnode) => {
-      const { proposal } = vnode.attrs;
-      if (!proposal || !proposal.threadId) return;
-      const path = getProposalUrlPath(
-        ProposalType.OffchainThread,
-        `${proposal.threadId}`,
-        false,
-        proposal['chain']
-      );
-      return m('.ProposalHeaderThreadLink', [
-        link('a.thread-link', path, [
-          'Go to discussion',
-          m(CWIcon, { iconName: 'externalLink', iconSize: 'small' }),
-        ]),
-      ]);
-    },
-  };
-
-export const ProposalHeaderSnapshotThreadLink: m.Component<{
-  thread: { id: string; title: string };
-}> = {
-  view: (vnode) => {
-    const { id, title } = vnode.attrs.thread;
-    if (!id) return;
-    const proposalLink = getProposalUrlPath(ProposalType.OffchainThread, id);
-
-    return m('.ProposalHeaderThreadLink', [
-      link('a.thread-link', proposalLink, [
-        decodeURIComponent(title),
-        m(CWIcon, { iconName: 'externalLink', iconSize: 'small' }),
-      ]),
-    ]);
-  },
-};
-
-export const ProposalHeaderSpacer: m.Component<{}> = {
-  view: (vnode) => {
+export const ProposalHeaderSpacer: m.Component = {
+  view: () => {
     return m('.ProposalHeaderSpacer', m.trust('&middot;'));
   },
 };
@@ -152,8 +37,6 @@ export const ProposalHeaderTopics: m.Component<{
     if (!proposal) return;
     if (!(proposal instanceof OffchainThread)) return;
     if (!proposal.topic) return;
-
-    const topicColor = '#72b483';
 
     return m('.ProposalHeaderTopics', [
       link(
@@ -249,28 +132,6 @@ export const ProposalHeaderViewCount: m.Component<{ viewCount: number }> = {
     return m('.ViewCountBlock', pluralize(viewCount, 'view'));
   },
 };
-
-// export const ProposalHeaderLinkThreadsMenuItem: m.Component<
-//   {
-//     item: OffchainThread;
-//   },
-//   {}
-// > = {
-//   view: (vnode) => {
-//     const { item } = vnode.attrs;
-//     return m(MenuItem, {
-//       label: 'Link offchain thread',
-//       class: 'link-offchain-thread',
-//       onclick: async (e) => {
-//         e.preventDefault();
-//         app.modals.create({
-//           modal: LinkedThreadModal,
-//           data: { linkingProposal: item },
-//         });
-//       },
-//     });
-//   },
-// };
 
 export const ProposalTitleEditMenuItem: m.Component<{
   item: AnyProposal;
