+++ client/scripts/views/pages/view_proposal/index.ts
@@ -69,10 +69,6 @@ import { modelFromServer as modelReactionCountFromServer } from 'controllers/ser
 import { SnapshotProposal } from 'helpers/snapshot_utils';
 import OffchainPoll from 'client/scripts/models/OffchainPoll';
 import {
-  ProposalHeaderExternalLink,
-  ProposalHeaderBlockExplorerLink,
-  ProposalHeaderVotingInterfaceLink,
-  ProposalHeaderThreadLink,
   ProposalHeaderTopics,
   ProposalHeaderTitle,
   ProposalHeaderStage,
@@ -120,6 +116,12 @@ import { LinkedThreadsCard } from './linked_threads_card';
 import { CommentReactionButton } from '../../components/reaction_button/comment_reaction_button';
 import { ThreadReactionButton } from '../../components/reaction_button/thread_reaction_button';
 import { ProposalPoll } from './poll';
+import {
+  ProposalHeaderExternalLink,
+  ProposalHeaderThreadLink,
+  ProposalHeaderBlockExplorerLink,
+  ProposalHeaderVotingInterfaceLink,
+} from './proposal_header_links';
 
 const MAX_THREAD_LEVEL = 2;
 
@@ -435,12 +437,12 @@ const ProposalHeader: m.Component<
             m('.proposal-body-link', [
               proposal instanceof OffchainThread &&
                 proposal.kind === OffchainThreadKind.Link && [
-                  !vnode.state.editing
-                    ? m(ProposalHeaderExternalLink, { proposal })
-                    : m(ProposalLinkEditor, {
+                  vnode.state.editing
+                    ? m(ProposalLinkEditor, {
                         item: proposal,
                         parentState: vnode.state,
-                      }),
+                      })
+                    : m(ProposalHeaderExternalLink, { proposal }),
                 ],
               !(proposal instanceof OffchainThread) &&
                 (proposal['blockExplorerLink'] ||
