+++ client/styles/pages/view_proposal/index.scss
@@ -125,6 +125,13 @@ $discussion-title-font-size: 1.45rem;
         width: 70%;
         padding-bottom: 10px;
       }
+      .proposal-body-link {
+        display: flex;
+
+        > * div {
+          margin-right: 8px;
+        }
+      }
       .proposal-body-meta {
         display: flex;
         flex-direction: row;
@@ -216,38 +223,6 @@ $discussion-title-font-size: 1.45rem;
             margin-right: 3px;
           }
         }
-      }
-    }
-    .ProposalHeaderExternalLink,
-    .ProposalHeaderBlockExplorerLink,
-    .ProposalHeaderVotingInterfaceLink,
-    .ProposalHeaderThreadLink {
-      display: inline-block;
-      margin-right: 12px;
-      margin-top: 12px;
-      font-weight: 600;
-      a:link,
-      a:visited {
-        display: block;
-        padding: 8px 16px;
-        border-radius: 99px;
-        background: $neutral-200;
-        color: $neutral-500;
-        text-decoration: none !important;
-      }
-      svg {
-        stroke: $primary-500;
-        position: relative;
-        top: -1px;
-        margin-left: 5px;
-        $size: 14px;
-        height: $size;
-        width: $size;
-        min-height: $size;
-        min-width: $size;
-      }
-      a:hover {
-        background: $neutral-300;
       }
     }
   }
