+++ client/styles/pages/view_proposal/proposal_header_links.scss
@@ -1,0 +1,24 @@
+@import 'shared';
+
+.ProposalHeaderLink {
+  a:link,
+  a:visited {
+    align-items: center;
+    background: $neutral-200;
+    border-radius: 99px;
+    color: $neutral-500;
+    display: flex;
+    font-weight: 600;
+    padding: 8px 16px;
+    text-decoration: none;
+  }
+
+  svg {
+    color: $primary-500;
+    margin-left: 8px;
+  }
+
+  a:hover {
+    background: $neutral-300;
+  }
+}
